+++ Data/ExportedCalendarEvents.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2022-07-12 17-19-04 ]] --
+-- [[ Exported at 2022-07-27 09-57-55 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --